backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-ollama
langchain-google-genai
python-dotenv